Move MetaFrame class into a separate file
